Adds node dev dependency
