+++ package.json
@@ -45,6 +45,7 @@
   },
   "dependencies": {},
   "devDependencies": {
+    "@types/node": "^20.2.5",
     "baldrick-zest-engine": "^0.9.0",
     "baldrick-zest-mess": "^0.17.0",
     "ts-node": "^10.9.1",
