+++ test/index.test.ts
@@ -1,5 +1,4 @@
-
-import { test } from 'node:test';
+import {test} from 'node:test';
 
 test('no test yet', () => {
   console.log('no test yet');
